backend/requirements.txt
fastapi
uvicorn
python-multipart
replicate
httpx
Pillow
python-dotenv
requests